moved layout-components to subcategory
